RW-599: Removed dependency on variable-time ConcurrentLinkedQueue.size() operation.
Updated implementation of RBroker, replacing ConcurrentLinkedQueue.size() call with isEmpty() call to reduce runtime overhead.

Per the JDK documentation on ConcurrentLinkedQueue:

"Beware that, unlike in most collections, the size method is NOT a constant-time operation. Because of the asynchronous nature of these queues, determining the current number of elements requires an O(n) traversal. Additionally, if elements are added or removed during execution of this method, the returned result may be inaccurate. Thus, this method is typically not very useful in concurrent applications."